Gravity-controller
I'm a student of methods to extend our knowledge of physical sciences where our understanding is incomplete. My current focus is regarding the greater potential that can be derived from the effects of electromagnetism. This includes both the theoretical & practical potential options regarding electromagnetism. I believe there is much that has yet to be understood in this science.